First review

+++ yatube/posts/views.py
@@ -164,8 +164,15 @@ def profile_follow(request, username):
 
 @login_required
 def profile_unfollow(request, username):
-    Follow.objects.get(
-        user=request.user,
-        author=get_object_or_404(User, username=username)
-    ).delete()
+    if (
+        request.user != get_object_or_404(User, username=username)
+        and Follow.objects.filter(
+            user=request.user,
+            author=get_object_or_404(User, username=username)
+        ).exists()
+    ):
+        Follow.objects.get(
+            user=request.user,
+            author=get_object_or_404(User, username=username)
+        ).delete()
     return redirect('posts:profile', username=username)
